backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.3
SQLAlchemy==2.0.31
psycopg[binary]==3.1.19
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.3.4
httpx==0.27.0
email-validator==2.2.0
